Move BaseParserTest
