+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/ast/ASTBooleanLiteralTest.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 import org.junit.Test;
 
-import net.sourceforge.pmd.lang.java.symboltable.BaseParserTest;
+import net.sourceforge.pmd.lang.java.BaseParserTest;
 
 public class ASTBooleanLiteralTest extends BaseParserTest {
