+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/ast/ASTSwitchStatementTest.java
@@ -7,7 +7,7 @@ package net.sourceforge.pmd.lang.java.ast;
 import org.junit.Assert;
 import org.junit.Test;
 
-import net.sourceforge.pmd.lang.java.symboltable.BaseParserTest;
+import net.sourceforge.pmd.lang.java.BaseParserTest;
 
 public class ASTSwitchStatementTest extends BaseParserTest {
