+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/ast/CommentAssignmentTest.java
@@ -13,7 +13,7 @@ import java.util.List;
 import org.junit.Assert;
 import org.junit.Test;
 
-import net.sourceforge.pmd.lang.java.symboltable.BaseParserTest;
+import net.sourceforge.pmd.lang.java.BaseParserTest;
 
 public class CommentAssignmentTest extends BaseParserTest {
