+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/ast/FormalCommentTest.java
@@ -10,7 +10,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import net.sourceforge.pmd.lang.ast.impl.javacc.JavaccToken;
-import net.sourceforge.pmd.lang.java.symboltable.BaseParserTest;
+import net.sourceforge.pmd.lang.java.BaseParserTest;
 
 public class FormalCommentTest extends BaseParserTest {
