+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/symboltable/MethodScopeTest.java
@@ -13,6 +13,7 @@ import org.junit.Ignore;
 import org.junit.Test;
 
 import net.sourceforge.pmd.PMD;
+import net.sourceforge.pmd.lang.java.BaseParserTest;
 import net.sourceforge.pmd.lang.java.ast.ASTCompilationUnit;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
 import net.sourceforge.pmd.lang.symboltable.NameDeclaration;
